add touch between rows
