+++ src/platform/packages/private/kbn-grid-layout/grid/grid_panel/drag_handle/default_drag_handle.tsx
@@ -1,0 +1,67 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+import React from 'react';
+import { EuiIcon, useEuiTheme } from '@elastic/eui';
+import { css } from '@emotion/react';
+import { i18n } from '@kbn/i18n';
+import { UserMouseEvent, UserTouchEvent } from '../../types';
+
+export const DefaultDragHandle = ({
+  onDragStart,
+}: {
+  onDragStart: (e: UserMouseEvent | UserTouchEvent) => void;
+}) => {
+  const { euiTheme } = useEuiTheme();
+
+  return (
+    <button
+      onMouseDown={onDragStart}
+      onTouchStart={onDragStart}
+      aria-label={i18n.translate('kbnGridLayout.dragHandle.ariaLabel', {
+        defaultMessage: 'Drag to move',
+      })}
+      className="kbnGridPanel__dragHandle"
+      css={css`
+        opacity: 0;
+        display: flex;
+        cursor: move;
+        position: absolute;
+        align-items: center;
+        justify-content: center;
+        top: -${euiTheme.size.l};
+        width: ${euiTheme.size.l};
+        height: ${euiTheme.size.l};
+        z-index: ${euiTheme.levels.modal};
+        margin-left: ${euiTheme.size.s};
+        border: 1px solid ${euiTheme.border.color};
+        border-bottom: none;
+        background-color: ${euiTheme.colors.backgroundBasePlain};
+        border-radius: ${euiTheme.border.radius} ${euiTheme.border.radius} 0 0;
+        cursor: grab;
+        transition: ${euiTheme.animation.slow} opacity;
+        .kbnGridPanel:hover &,
+        .kbnGridPanel:focus-within &,
+        &:active,
+        &:focus {
+          opacity: 1 !important;
+        }
+        &:active {
+          cursor: grabbing;
+        }
+        .kbnGrid--static &,
+        .kbnGridPanel--expanded & {
+          display: none;
+        }
+        touch-action: none;
+      `}
+    >
+      <EuiIcon type="grabOmnidirectional" />
+    </button>
+  );
+};
