+++ src/platform/packages/private/kbn-grid-layout/grid/grid_panel/grid_panel.tsx
@@ -13,7 +13,7 @@ import { combineLatest, skip } from 'rxjs';
 import { useEuiTheme } from '@elastic/eui';
 import { css } from '@emotion/react';
 
-import { GridLayoutStateManager, PanelInteractionEvent, UserInteractionEvent } from '../types';
+import { GridLayoutStateManager } from '../types';
 import { DragHandle, DragHandleApi } from './drag_handle';
 import { ResizeHandle } from './resize_handle';
 
@@ -24,46 +24,13 @@ export interface GridPanelProps {
     panelId: string,
     setDragHandles?: (refs: Array<HTMLElement | null>) => void
   ) => React.ReactNode;
-  interactionStart: (type: PanelInteractionEvent['type'] | 'drop', e: UserInteractionEvent) => void;
   gridLayoutStateManager: GridLayoutStateManager;
 }
 
 export const GridPanel = forwardRef<HTMLDivElement, GridPanelProps>(
-  (
-    { panelId, rowIndex, renderPanelContents, interactionStart, gridLayoutStateManager },
-    panelRef
-  ) => {
+  ({ panelId, rowIndex, renderPanelContents, gridLayoutStateManager }, panelRef) => {
     const [dragHandleApi, setDragHandleApi] = useState<DragHandleApi | null>(null);
     const { euiTheme } = useEuiTheme();
-
-    useEffect(() => {
-      const onDropEventHandler = (dropEvent: MouseEvent) => interactionStart('drop', dropEvent);
-      /**
-       * Subscription to add a singular "drop" event handler whenever an interaction starts -
-       * this is handled in a subscription so that it is not lost when the component gets remounted
-       * (which happens when a panel gets dragged from one grid row to another)
-       */
-      const dropEventSubscription = gridLayoutStateManager.interactionEvent$.subscribe((event) => {
-        if (!event || event.id !== panelId) return;
-
-        /**
-         * By adding the "drop" event listener to the document rather than the drag/resize event handler,
-         * we prevent the element from getting "stuck" in an interaction; however, we only attach this event
-         * listener **when the drag/resize event starts**, and it only executes once (i.e. it removes itself
-         * once it executes, so we don't have to manually remove it outside of the unmount condition)
-         */
-        document.addEventListener('mouseup', onDropEventHandler, {
-          once: true,
-          passive: true,
-        });
-      });
-
-      return () => {
-        dropEventSubscription.unsubscribe();
-        document.removeEventListener('mouseup', onDropEventHandler); // removes the event listener on row change
-      };
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
 
     /** Set initial styles based on state at mount to prevent styles from "blipping" */
     const initialStyles = useMemo(() => {
@@ -195,10 +162,15 @@ export const GridPanel = forwardRef<HTMLDivElement, GridPanelProps>(
         <DragHandle
           ref={setDragHandleApi}
           gridLayoutStateManager={gridLayoutStateManager}
-          interactionStart={interactionStart}
+          panelId={panelId}
+          rowIndex={rowIndex}
         />
         {panelContents}
-        <ResizeHandle interactionStart={interactionStart} />
+        <ResizeHandle
+          gridLayoutStateManager={gridLayoutStateManager}
+          panelId={panelId}
+          rowIndex={rowIndex}
+        />
       </div>
     );
   }
