+++ src/platform/packages/private/kbn-grid-layout/grid/grid_row/grid_row.tsx
@@ -16,9 +16,8 @@ import { css } from '@emotion/react';
 
 import { DragPreview } from '../drag_preview';
 import { GridPanel } from '../grid_panel';
-import { GridLayoutStateManager, PanelInteractionEvent, UserInteractionEvent } from '../types';
+import { GridLayoutStateManager } from '../types';
 import { getKeysInOrder } from '../utils/resolve_grid_row';
-import { isMouseEvent, isTouchEvent } from '../utils/sensors';
 import { GridRowHeader } from './grid_row_header';
 
 export interface GridRowProps {
@@ -27,12 +26,11 @@ export interface GridRowProps {
     panelId: string,
     setDragHandles?: (refs: Array<HTMLElement | null>) => void
   ) => React.ReactNode;
-  setInteractionEvent: (interactionData?: PanelInteractionEvent) => void;
   gridLayoutStateManager: GridLayoutStateManager;
 }
 
 export const GridRow = forwardRef<HTMLDivElement, GridRowProps>(
-  ({ rowIndex, renderPanelContents, setInteractionEvent, gridLayoutStateManager }, gridRef) => {
+  ({ rowIndex, renderPanelContents, gridLayoutStateManager }, gridRef) => {
     const currentRow = gridLayoutStateManager.gridLayout$.value[rowIndex];
 
     const [panelIds, setPanelIds] = useState<string[]>(Object.keys(currentRow.panels));
@@ -63,11 +61,11 @@ export const GridRow = forwardRef<HTMLDivElement, GridRowProps>(
         /** Update the styles of the grid row via a subscription to prevent re-renders */
         const interactionStyleSubscription = combineLatest([
           gridLayoutStateManager.interactionEvent$,
-          gridLayoutStateManager.gridLayout$,
           gridLayoutStateManager.runtimeSettings$,
+          gridLayoutStateManager.gridLayout$,
         ])
           .pipe(skip(1)) // skip the first emit because the `initialStyles` will take care of it
-          .subscribe(([interactionEvent, gridLayout, runtimeSettings]) => {
+          .subscribe(([interactionEvent, runtimeSettings]) => {
             const rowRef = gridLayoutStateManager.rowRefs.current[rowIndex];
             if (!rowRef) return;
 
@@ -133,8 +131,23 @@ export const GridRow = forwardRef<HTMLDivElement, GridRowProps>(
             }
           });
 
+        /**
+         * Ensure the row re-renders to reflect the new panel order after a drag-and-drop interaction, since
+         * the order of rendered panels need to be aligned with how they are displayed in the grid for accessibility
+         * reasons (screen readers and focus management).
+         */
+        const stableGridLayoutSubscription = gridLayoutStateManager.stableGridLayout$.subscribe(
+          (stableGridLayout) => {
+            const newPanelIdsInOrder = getKeysInOrder(stableGridLayout[rowIndex].panels);
+            if (panelIdsInOrder.join() !== newPanelIdsInOrder.join()) {
+              setPanelIdsInOrder(newPanelIdsInOrder);
+            }
+          }
+        );
+
         return () => {
           interactionStyleSubscription.unsubscribe();
+          stableGridLayoutSubscription.unsubscribe();
           rowStateSubscription.unsubscribe();
         };
       },
@@ -156,39 +169,6 @@ export const GridRow = forwardRef<HTMLDivElement, GridRowProps>(
               rowIndex={rowIndex}
               gridLayoutStateManager={gridLayoutStateManager}
               renderPanelContents={renderPanelContents}
-              interactionStart={(type, e) => {
-                e.stopPropagation();
-
-                // Disable interactions when a panel is expanded
-                const isInteractive = gridLayoutStateManager.expandedPanelId$.value === undefined;
-                if (!isInteractive) return;
-
-                const panelRef = gridLayoutStateManager.panelRefs.current[rowIndex][panelId];
-                if (!panelRef) return;
-
-                if (type === 'drop') {
-                  setInteractionEvent(undefined);
-                  /**
-                   * Ensure the row re-renders to reflect the new panel order after a drag-and-drop interaction, since
-                   * the order of rendered panels need to be aligned with how they are displayed in the grid for accessibility
-                   * reasons (screen readers and focus management).
-                   */
-                  setPanelIdsInOrder(
-                    getKeysInOrder(gridLayoutStateManager.gridLayout$.getValue()[rowIndex].panels)
-                  );
-                } else {
-                  const panelRect = panelRef.getBoundingClientRect();
-                  const pointerOffsets = getPointerOffsets(e, panelRect);
-
-                  setInteractionEvent({
-                    type,
-                    id: panelId,
-                    panelDiv: panelRef,
-                    targetRowIndex: rowIndex,
-                    pointerOffsets,
-                  });
-                }
-              }}
               ref={(element) => {
                 if (!gridLayoutStateManager.panelRefs.current[rowIndex]) {
                   gridLayoutStateManager.panelRefs.current[rowIndex] = {};
@@ -200,7 +180,7 @@ export const GridRow = forwardRef<HTMLDivElement, GridRowProps>(
         }),
         {}
       );
-    }, [panelIds, gridLayoutStateManager, renderPanelContents, rowIndex, setInteractionEvent]);
+    }, [panelIds, gridLayoutStateManager, renderPanelContents, rowIndex]);
 
     return (
       <div
@@ -242,32 +222,3 @@ export const GridRow = forwardRef<HTMLDivElement, GridRowProps>(
     );
   }
 );
-
-const defaultPointerOffsets = {
-  top: 0,
-  left: 0,
-  right: 0,
-  bottom: 0,
-};
-
-function getPointerOffsets(e: UserInteractionEvent, panelRect: DOMRect) {
-  if (isTouchEvent(e)) {
-    if (e.touches.length > 1) return defaultPointerOffsets;
-    const touch = e.touches[0];
-    return {
-      top: touch.clientY - panelRect.top,
-      left: touch.clientX - panelRect.left,
-      right: touch.clientX - panelRect.right,
-      bottom: touch.clientY - panelRect.bottom,
-    };
-  }
-  if (isMouseEvent(e)) {
-    return {
-      top: e.clientY - panelRect.top,
-      left: e.clientX - panelRect.left,
-      right: e.clientX - panelRect.right,
-      bottom: e.clientY - panelRect.bottom,
-    };
-  }
-  throw new Error('Invalid event type');
-}
